requirements.txt
httpx>=0.24.0  
rich>=13.4.2  
PyYAML>=6.0  
tenacity>=8.2.2  
numpy>=1.24.0  